+++ src/client/containers/ProfilePage/NewBoard.component.js
@@ -1,4 +1,5 @@
 import React from 'react';
+// eslint-disable-next-line import/no-extraneous-dependencies
 import { useHistory } from 'react-router';
 import './NewBoard.styles.css';
